feat: add substring function and move vm functions to its own file
