+++ src/vm/mod.rs
@@ -1,3 +1,5 @@
 pub(crate) mod expression;
 pub(crate) mod planner;
 pub(crate) mod statement;
+
+mod functions;
